+++ test/Fixture/Closures/NoNullableReturnTypeDeclarationRule/Success/closure-without-return-type-declaration.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Closures\NoNullableReturnTypeDeclarationRule\Success;
+
+$foo = function () {
+    return 'Hello';
+};