+++ test/Fixture/Functions/NoNullableReturnTypeDeclarationRule/Success/anonymous-function-without-return-type-declaration.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Localheinz\PHPStan\Rules\Test\Fixture\Functions\NoNullableReturnTypeDeclarationRule\Success;
-
-$foo = function () {
-    return 'Hello';
-};