+++ CHANGELOG.md
@@ -12,10 +12,12 @@ For a full diff see [`0.2.0...master`](https://github.com/localheinz/phpstan-rul
 ### Added
 
 * added `Functions\NoNullableReturnTypeDeclarationRule`, which reports an
-  error when a named function has a nullable return type declaration, and
+  error when a function has a nullable return type declaration, and
   `Methods\NoNullableReturnTypeDeclarationRule`, which reports an error
   when a method declared on an anonymous class, a class, or an interface has a
   nullable return type declaration ([#16](https://github.com/localheinz/phpstan-rules/pull/16)), by [@localheinz](https://github.com/localheinz)
+* added `Closures\NoNullableReturnTypeDeclarationRule`, which reports an
+  error when a closure has a nullable return type declaration ([#28](https://github.com/localheinz/phpstan-rules/pull/28)), by [@localheinz](https://github.com/localheinz)
 
 ## [`0.2.0`](https://github.com/localheinz/phpstan-rules/releases/tag/0.2.0)